test(sample): `(foo + bar) as u8`

+++ test/sample.rs
@@ -444,7 +444,7 @@ trait TraitWithBlockComment<'a> {
 		let refchar = &'c';
 
 		let math = 1+2.0 - 6%var / CONST * 9_00;
-		let bit = 1 & &foo ^ 2 | 10 & (1 + 2 / 3);
+		let bit = 1 & &foo ^ 2 | 10 & (1 + 2 / 3) as u8;
 		let compare = 1 > 2 && 1 == 1 ||
 			2 >= 2 &&
 			2 <= 2
